requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
facebook-sdk>=3.1.0
python-dotenv>=1.1.0
loguru>=0.7.0
pydantic>=2.0.0
fastapi_mcp
python-multipart
